feat!: change the parser repo to an installable plugin
First, makes this repo installable via plugin manager. In addition it
already takes care of most setup for the user:
- file type detection
- parser registration
- injections queries

Second, it defines `helm` as a language that uses the gotmpl language,
but with different injection queries. This allows the standard gotmpl
language to use simple text for its text nodes and only injects `yaml`
as a language in `helm` templates. The injections are also set to
combined, to fix some highlighting issues.

Third, it increases the priority of gotmpl highlight queries, so that
the gotmpl highlights are always visible above the highlights of any
injection language.